pyproject.toml
[project]
name = "moneymanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-ninja>=1.5.3",
    "pyjwt>=2.8.0",
    "email-validator>=2.1.0",
    "pydantic[email]>=2.12.5",
    "django-stubs>=5.2.9",
]
